requirements.txt
asgiref==3.7.2
channels==4.0.0
crispy-bootstrap4==2023.1
daphne==4.0.0
Django==5.0.6
django-crispy-forms==2.1
djangorestframework==3.14.0
pillow==10.2.0
pyasn1==0.5.1
pycparser==2.21
sqlparse==0.4.4
typing_extensions==4.9.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-environ
